requirements.txt
aiofiles==0.8.0
aniso8601==9.0.1
anyio==3.5.0
asgiref==3.5.0
async-exit-stack==1.0.1
async-generator==1.10
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.58.1
google-auth==2.6.0
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.5
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
inflect==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
orjson==3.6.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.31
starlette==0.13.4
toml==0.10.2
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.8
uvicorn==0.17.1
watchgod==0.7
websockets==10.1
